Updated the ReadMe and fixed leftover code issues
-- Fixed multiple minor code issues
-- Updated the ReadMe to match the current codebase state.

+++ src/sl_experiment/shared_components/module_interfaces.py
@@ -993,7 +993,7 @@ class ValveInterface(ModuleInterface):
                 f"{_MAXIMUM_VALVE_PULSE_DURATION_MS} ms."
             )
             console.echo(message=message, level=LogLevel.WARNING)
-            pulse_duration = _MAXIMUM_VALVE_PULSE_DURATION_US
+            pulse_duration = np.float64(_MAXIMUM_VALVE_PULSE_DURATION_US)
 
         return np.uint32(np.round(pulse_duration))
